fold everything into a single class, which is now the `SapiEmitter` (since there is only one emitter)
